+++ src/block.rs
@@ -34,6 +34,7 @@ pub struct BlockLayout {
     cross_axis_alignment: AxisAlignment,
     child: Box<dyn Layout>,
     errors: Vec<LayoutError>,
+    label: Option<String>,
 }
 
 impl Default for BlockLayout {
@@ -49,6 +50,7 @@ impl Default for BlockLayout {
             cross_axis_alignment: AxisAlignment::default(),
             errors: vec![],
             child: Box::new(EmptyLayout::default()),
+            label: None,
         }
     }
 }
@@ -67,6 +69,11 @@ impl BlockLayout {
 
     pub fn set_id(mut self, id: GlobalId) -> Self {
         self.id = id;
+        self
+    }
+
+    pub fn with_label(mut self, label: &str) -> Self {
+        self.label = Some(label.to_string());
         self
     }
 
@@ -130,6 +137,10 @@ impl BlockLayout {
 }
 
 impl Layout for BlockLayout {
+    fn label(&self) -> String {
+        self.label.clone().unwrap_or("BlockLayout".to_string())
+    }
+
     fn id(&self) -> GlobalId {
         self.id
     }