+++ src/horizontal.rs
@@ -31,6 +31,7 @@ pub struct HorizontalLayout {
     cross_axis_alignment: AxisAlignment,
     children: Vec<Box<dyn Layout>>,
     errors: Vec<LayoutError>,
+    label: Option<String>,
 }
 
 impl HorizontalLayout {
@@ -55,6 +56,11 @@ impl HorizontalLayout {
     /// ```
     pub fn add_child(mut self, child: impl Layout + 'static) -> Self {
         self.children.push(Box::new(child));
+        self
+    }
+
+    pub fn with_label(mut self, label: &str) -> Self {
+        self.label = Some(label.to_string());
         self
     }
 
@@ -229,6 +235,10 @@ impl HorizontalLayout {
 }
 
 impl Layout for HorizontalLayout {
+    fn label(&self) -> String {
+        self.label.clone().unwrap_or("HorizontalLayout".to_string())
+    }
+
     fn id(&self) -> GlobalId {
         self.id
     }