+++ src/lib.rs
@@ -88,6 +88,8 @@
 #![warn(clippy::suspicious_operation_groupings)]
 #![warn(clippy::imprecise_flops)]
 pub mod block;
+#[cfg(feature = "debug-tools")]
+pub mod debug;
 pub mod empty;
 mod error;
 pub mod horizontal;
@@ -138,18 +140,9 @@ pub fn solve_layout(root: &mut dyn Layout, window_size: Size) -> Vec<LayoutError
 }
 
 /// A layout node.
-///
-/// ## Details
-///
-/// Each layout node has minimum and maximum constraints, these describe the bounds
-/// of the layout node i.e. the max and min space it's allowed to take up.
-///
-/// ### Axes
-/// Each node has two axes: the main axis and the cross axis. The main axis is the
-/// axis along which content flows and the cross axis is the axis perpendicular
-/// to the cross axis. For most nodes the main axis is the x-axis while the
-/// cross axis is the y-axis.
 pub trait Layout: Debug + private::Sealed {
+    fn label(&self) -> String;
+
     /// Solve the minimum constraints of each [`Layout`] node recursively
     fn solve_min_constraints(&mut self) -> (f32, f32);